test(geometry): pin the station-velocity dtype for CI's numpy stubs
`np.zeros(n)` with no dtype is one of the three shapes that have failed on the
CI runners' own numpy. Explicit `dtype=np.float64`, no behaviour change.

Negative controls run against this commit's tests (mutate, run, `git checkout --`):
- transport term dropped          -> 5 tests fail
- range-rate sign flipped         -> the same 5 fail
- transport term sign reversed    -> the same 5 fail
- rotating-frame formulation
  (`- omega x A r_body` instead)  -> all 45 pass, correctly: the two differ by
  `omega x rho`, which is orthogonal to `rho` and drops out of the projection.

+++ src/orbital_engine/geometry.py
@@ -379,7 +379,9 @@ def elevation_azimuth(
         )
         # omega x r_station, expressed in the body-fixed frame: rotation is about +z.
         station_vel = float(omega) * np.stack(
-            [-station_fixed[:, 1], station_fixed[:, 0], np.zeros(station_fixed.shape[0])], axis=-1
+            [-station_fixed[:, 1], station_fixed[:, 0],
+             np.zeros(station_fixed.shape[0], dtype=np.float64)],
+            axis=-1,
         )
         rho_dot = v_rot[..., 0][:, np.newaxis, :, :] - station_vel[np.newaxis, :, np.newaxis, :]
         # |rho| = 0 forces the numerator to 0 as well, so dividing by 1.0 there returns exactly 0.0.
